refactor: performance & style
