+++ src/domain/cpu/index.ts
@@ -4,7 +4,7 @@ import { CpuScheduler } from "./CpuScheduler";
 (async () => {
   const scheduler = new CpuScheduler()
     .setWorkerCount(4)
-    .generateDataSet(100_000_000, 7);
+    .generateDataSet(1_000_000, 7);
   await scheduler.ready();
 
   for (const i of Array.from(Array(100).keys())) {
